# backend/openapi.yaml
openapi: 3.1.0
info:
    title: VidyaSetu
    version: 0.1.0
    description: |
        VidyaSetu is a role-based education and workshop management backend API designed for Admin, Institutional Admin, Educator, Student, and Public users.

        The system supports the complete academic lifecycle including authentication, content delivery, assessments, certification, analytics, communication, and AI-powered features.

        #### Core capabilities include:

        - **Secure JWT-based authentication** and role-based access control for admin, institutional admin, educator, and student users.
        - **Institution and workshop administration**, including workshop creation, updates, enrollment tracking, assignments, and lifecycle management.
        - **Centralized content management** for modules, study materials, and learning resources uploaded by educators and accessed by students.
        - **Assessment management** for quizzes, tests, and assignments, including question creation, student attempts, answer submission, and result storage.
        - **Automated evaluation support** for objective assessments to reduce manual grading effort and improve consistency.
        - **Certificate generation** for successful workshop completion and public certificate verification using unique verification codes or identifiers.
        - **Performance tracking and reporting** for student progress, workshop statistics, assessment outcomes, completion rates, and trend analysis.
...

...

...

...

...

...

...

...

...
components:
    schemas:
        AIGenerationStatus:
            enum:
                - pending
                - processing
                - completed
                - failed
            title: AIGenerationStatus
            type: string
        AdminAIReportCreateRequest:
            properties:
                date_from:
                    anyOf:
                        - format: date
                          type: string
                        - type: "null"
                    title: Date From
                date_to:
                    anyOf:
                        - format: date
                          type: string
                        - type: "null"
                    title: Date To
                focus_areas:
                    items:
                        type: string
                    title: Focus Areas
                    type: array
                force_regenerate:
                    default: false
                    title: Force Regenerate
                    type: boolean
                institution_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Institution Id
            title: AdminAIReportCreateRequest
            type: object
        AdminAIReportCreateResponse:
            properties:
                deduplicated:
                    title: Deduplicated
                    type: boolean
                from_cache:
                    title: From Cache
                    type: boolean
                report_id:
                    title: Report Id
                    type: string
                status:
                    $ref: "#/components/schemas/AIGenerationStatus"
            required:
                - report_id
                - status
                - from_cache
                - deduplicated
            title: AdminAIReportCreateResponse
            type: object
        AdminAIReportListItem:
            properties:
                created_at:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: Created At
                institution_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Institution Id
                model_name:
                    title: Model Name
                    type: string
                prompt_version:
                    title: Prompt Version
                    type: string
                report_id:
                    title: Report Id
                    type: string
                source_entity_id:
                    title: Source Entity Id
                    type: string
                source_entity_type:
                    title: Source Entity Type
                    type: string
                status:
                    $ref: "#/components/schemas/AIGenerationStatus"
                summary_preview:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Summary Preview
                updated_at:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: Updated At
            required:
                - report_id
                - status
                - source_entity_type
                - source_entity_id
                - prompt_version
                - model_name
            title: AdminAIReportListItem
            type: object
        AdminAIReportResultResponse:
            properties:
                created_at:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: Created At
                model_name:
                    title: Model Name
                    type: string
                prompt_version:
                    title: Prompt Version
                    type: string
                report_id:
                    title: Report Id
                    type: string
                result:
                    $ref: "#/components/schemas/AdminReportStructuredOutput"
                status:
                    $ref: "#/components/schemas/AIGenerationStatus"
                updated_at:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: Updated At
            required:
                - report_id
                - status
                - result
                - prompt_version
                - model_name
            title: AdminAIReportResultResponse
            type: object
        AdminAIReportStatusResponse:
            properties:
                created_at:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: Created At
                error_details:
                    anyOf:
                        - additionalProperties: true
                          type: object
                        - type: "null"
                    title: Error Details
                report_id:
                    title: Report Id
                    type: string
                status:
                    $ref: "#/components/schemas/AIGenerationStatus"
                updated_at:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: Updated At
            required:
                - report_id
                - status
            title: AdminAIReportStatusResponse
            type: object
...
        AdminReportDataWindow:
            properties:
                end_date:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: End Date
                scope:
                    title: Scope
                    type: string
                start_date:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Start Date
            required:
                - scope
            title: AdminReportDataWindow
            type: object
        AdminReportRecommendationItem:
            properties:
                action:
                    title: Action
                    type: string
                priority:
                    description: low|medium|high
                    title: Priority
                    type: string
                rationale:
                    title: Rationale
                    type: string
                title:
                    title: Title
                    type: string
            required:
                - title
                - action
                - rationale
                - priority
            title: AdminReportRecommendationItem
            type: object
        AdminReportStructuredOutput:
            properties:
                caveats:
                    items:
                        type: string
                    title: Caveats
                    type: array
                data_window:
                    $ref: "#/components/schemas/AdminReportDataWindow"
                key_insights:
                    items:
                        type: string
                    title: Key Insights
                    type: array
                recommendations:
                    items:
                        $ref: "#/components/schemas/AdminReportRecommendationItem"
                    title: Recommendations
                    type: array
                risk_flags:
                    items:
                        type: string
                    title: Risk Flags
                    type: array
                summary:
                    title: Summary
                    type: string
                trend_highlights:
                    items:
                        type: string
                    title: Trend Highlights
                    type: array
            required:
                - summary
                - data_window
            title: AdminReportStructuredOutput
            type: object
...
        HTTPValidationError:
            properties:
                detail:
                    items:
                        $ref: "#/components/schemas/ValidationError"
                    title: Detail
                    type: array
            title: HTTPValidationError
            type: object
...
        Page_AdminAIReportListItem_:
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/AdminAIReportListItem"
                    title: Items
                    type: array
                limit:
                    title: Limit
                    type: integer
                offset:
                    title: Offset
                    type: integer
                total:
                    title: Total
                    type: integer
            required:
                - items
                - total
                - offset
                - limit
            title: Page[AdminAIReportListItem]
            type: object
...
        ValidationError:
            properties:
                ctx:
                    title: Context
                    type: object
                input:
                    title: Input
                loc:
                    items:
                        anyOf:
                            - type: string
                            - type: integer
                    title: Location
                    type: array
                msg:
                    title: Message
                    type: string
                type:
                    title: Error Type
                    type: string
            required:
                - loc
                - msg
                - type
            title: ValidationError
            type: object
...
    securitySchemes:
        OAuth2PasswordBearer:
            flows:
                password:
                    scopes: {}
                    tokenUrl: /api/v1/auth/login
            type: oauth2
paths:
...
    /api/v1/ai/reports/:
        get:
            operationId: list_admin_ai_reports_api_v1_ai_reports__get
            parameters:
                - in: query
                  name: institution_id
                  required: false
                  schema:
                      anyOf:
                          - type: string
                          - type: "null"
                      title: Institution Id
                - description: Number of records to skip
                  in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      description: Number of records to skip
                      minimum: 0
                      title: Offset
                      type: integer
                - description: Maximum records to return
                  in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      description: Maximum records to return
                      maximum: 1000
                      minimum: 1
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page_AdminAIReportListItem_"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: List previous admin AI reports
            tags:
                - ai-reports
        post:
            operationId: create_admin_ai_report_api_v1_ai_reports__post
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/AdminAIReportCreateRequest"
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AdminAIReportCreateResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Create and generate an admin AI report
            tags:
                - ai-reports
    /api/v1/ai/reports/{report_id}/result:
        get:
            operationId: get_admin_ai_report_result_api_v1_ai_reports__report_id__result_get
            parameters:
                - in: path
                  name: report_id
                  required: true
                  schema:
                      title: Report Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AdminAIReportResultResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Get completed admin AI report result
            tags:
                - ai-reports
    /api/v1/ai/reports/{report_id}/status:
        get:
            operationId: get_admin_ai_report_status_api_v1_ai_reports__report_id__status_get
            parameters:
                - in: path
                  name: report_id
                  required: true
                  schema:
                      title: Report Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AdminAIReportStatusResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Get admin AI report generation status
            tags:
                - ai-reports
...